Add some scripts and format files.

+++ contracts/interfaces/ICore.sol
@@ -2,26 +2,25 @@
 pragma solidity ^0.8.2;
 
 interface ICore {
-
     // Structures
 
     struct Node {
         bool isExecuted;
         bool isSuccess;
-        uint executionBlock;
+        uint256 executionBlock;
         address caller;
     }
 
     struct Ticket {
-        uint  targetExecutionBlock;
-        uint  creationBlock;
+        uint256 targetExecutionBlock;
+        uint256 creationBlock;
         address tokenAddress;
         address caller;
         address from;
         address to;
         uint256 value;
         uint256 fee;
-        uint32  nonce;
+        uint32 nonce;
         bytes32 ticketHash;
         Node node;
     }
@@ -29,23 +28,21 @@ interface ICore {
     // Events
 
     event TicketCreate(
-        uint  targetExecutionBlock,
-        uint  creationBlock,
+        uint256 targetExecutionBlock,
+        uint256 creationBlock,
         address tokenAddress,
         address caller,
         address from,
         address to,
         uint256 value,
         uint256 fee,
-        uint32  nonce
+        uint32 nonce
     );
 
-    event TicketFail(
-        Ticket _ticket
-    );
+    event TicketFail(Ticket _ticket);
 
     function createTicket(
-        uint  targetExecutionBlock,
+        uint256 targetExecutionBlock,
         address _tokenAddress,
         address _from,
         address _to,
@@ -53,5 +50,8 @@ interface ICore {
         uint256 _fee
     ) external payable returns (bytes32);
 
-    function getTicket(bytes32 _ticketHash) external view returns (Ticket memory _ticket);
+    function getTicket(bytes32 _ticketHash)
+        external
+        view
+        returns (Ticket memory _ticket);
 }
